Add an option to add annotations to images.
Image annotations are represented in the same manner as regular
annotations as members of the annotations array. Their structure is
somewhat different though.

The new image annotations have a `type` property that differentiates
them from regular text annotations. They also have an `attachmentIndex`
property that tells which (0-indexed) image on the answer the annotation
is attached to.

Line annotations are represented by the following stucture that more or
less tries to mimic an SVG line element.

type Line = {
  type: 'line',
  attachmentIndex: number,
  x1: number,
  y1: number,
  x2: number,
  y2: number,
  message: string
}

Rectangle annotations are represented by the following structure. It
corresponds to the SVG rect element.

type Rect = {
  type: 'rect',
  attachmentIndex: number,
  x: number,
  y: number,
  width: number,
  height: number,
  message: string
}

All coordinates and widths are relative to the image dimensions. Point
(0, 0) is in the top left corner and (1, 1) in the bottom right corner
of the image.